@@ -68,40 +68,39 @@ class bdist_wheel(_bdist_wheel):
 # Setup configuration
 setup(
     name=_fetch_package_name(),
-    version=find_version("autowsgr/__init__.py"),
+    version=find_version('autowsgr/__init__.py'),
     packages=find_namespace_packages(
-        include=["autowsgr*", "awsg*"],
+        include=['autowsgr*', 'awsg*'],
         exclude=(
-            "docs",
-            "examples",
+            'docs',
+            'examples',
         ),
     ),
     include_package_data=True,
-    description="Auto Warship Girls Framework.",
+    description='Auto Warship Girls Framework.',
     long_description=_fetch_readme(),
-    long_description_content_type="text/markdown",
-    install_requires=_fetch_requirements("requirements.txt"),
-    setup_requires=["wheel"],
+    long_description_content_type='text/markdown',
+    install_requires=_fetch_requirements('requirements.txt'),
+    setup_requires=['wheel'],
     dependency_links=[
-        "https://download.pytorch.org/whl/cu123",
+        'https://download.pytorch.org/whl/cu123',
     ],
-    python_requires=">=3.9",
+    python_requires='>=3.10',
     package_data={
-        "": [
-            "version.txt",
-            "data/**",
-            "requirements.txt",
-            "bin/**",
-            "c_src/**",
+        '': [
+            'version.txt',
+            'data/**',
+            'requirements.txt',
+            'bin/**',
+            'c_src/**',
         ],  # 希望被打包的文件
     },
     classifiers=[
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
-        "Environment :: GPU :: NVIDIA CUDA",
-        "Topic :: Scientific/Engineering :: Artificial Intelligence",
-        "Topic :: System :: Distributed Computing",
+        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
+        'Environment :: GPU :: NVIDIA CUDA',
+        'Topic :: Scientific/Engineering :: Artificial Intelligence',
+        'Topic :: System :: Distributed Computing',
     ],
-    cmdclass={"bdist_wheel": bdist_wheel},
+    cmdclass={'bdist_wheel': bdist_wheel},
 )
