@@ -1,27 +1,24 @@
-from typing import Iterable
+from collections.abc import Iterable
+from typing import Any
 
 
-def remove_0_value_from_dict(d: dict):
-    rd = {}
-    for key, value in d.items():
-        if value != 0 and value != None:
-            rd[key] = value
-    return rd
+def remove_0_value_from_dict(d: dict) -> dict:
+    return {k: v for k, v in d.items() if v is not None and v != 0}
 
 
-def unzip_element(o: object):
+def unzip_element(o: object) -> list[object] | Any | list:
     if not isinstance(o, Iterable):
         return [o]
     res = []
     for value in o:
-        if isinstance(value, (list, set)):
+        if isinstance(value, list | set):
             res += unzip_element(value)
         else:
             res.append(value)
     return res
 
 
-def unorder_equal(o1: Iterable, o2: Iterable, skip=[None]):
+def unorder_equal(o1: Iterable, o2: Iterable, skip=None) -> bool:
     if skip is None:
-        skip = []
+        skip = [None]
     return (set(o1) - set(skip)) == (set(o2) - set(skip))
