@@ -1,4 +1,4 @@
-from .battle import BattlePlan
-from .decisive_battle import DecisiveBattle
-from .exercise import NormalExercisePlan
-from .normal_fight import NormalFightPlan
+from autowsgr.fight.battle import BattlePlan
+from autowsgr.fight.decisive_battle import DecisiveBattle
+from autowsgr.fight.exercise import NormalExercisePlan
+from autowsgr.fight.normal_fight import NormalFightPlan
