@@ -1,24 +1,19 @@
 import os
 import subprocess
-import time
+from typing import Literal
 
 from autowsgr.constants.custom_exceptions import ImageNotFoundErr
 from autowsgr.constants.data_roots import MAP_ROOT, TUNNEL_ROOT
 from autowsgr.constants.image_templates import IMG
 from autowsgr.fight.battle import BattleInfo, BattlePlan
 from autowsgr.fight.common import start_march
-from autowsgr.game.game_operation import DestroyShip, get_ship, quick_repair
+from autowsgr.game.game_operation import destroy_ship, get_ship, quick_repair
 from autowsgr.game.get_game_info import detect_ship_stats
-from autowsgr.ocr.ship_name import recognize
 from autowsgr.port.ship import Fleet, count_ship
 from autowsgr.timer import Timer
-from autowsgr.utils.api_image import (
-    crop_image,
-    crop_rectangle_relative,
-    cv2,
-    cv_show_image,
-)
+from autowsgr.utils.api_image import crop_image, crop_rectangle_relative, cv2
 from autowsgr.utils.io import count, yaml_to_dict
+
 
 """决战结构:
 上层控制+单点战斗
@@ -26,26 +21,25 @@ from autowsgr.utils.io import count, yaml_to_dict
 
 
 def is_ship(element):
-    return element not in ["长跑训练", "肌肉记忆", "黑科技"]
+    return element not in ['长跑训练', '肌肉记忆', '黑科技']
 
 
 def get_formation(fleet: Fleet, enemy: list):
-    anti_sub = count(["CL", "DD", "CVL"], enemy)
-    if fleet.exist("U-1206") and anti_sub <= 1:
-        return 4
-    elif anti_sub <= 0:
+    anti_sub = count(['CL', 'DD', 'CVL'], enemy)
+    if fleet.exist('U-1206') and anti_sub <= 1 or anti_sub <= 0:
         return 4
     return 2
 
 
 class DB:
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args, **kwargs) -> None:
         self.__dict__.update(kwargs)
 
     def make_decision(self, state):
         rule = self.__dict__
         if state in rule:
             return rule[state]
+        return None
 
 
 class DecisiveStats:
@@ -54,21 +48,21 @@ class DecisiveStats:
         self.timer = timer
         self.key_points = [
             [
-                "",
-            ]
+                '',
+            ],
         ]  # [chapter][map] (str)
         self.map_end = [
-            "",
+            '',
         ]  # [chapter] (str)
         self.enemy = [
             [
                 [
-                    "",
-                ]
-            ]
+                    '',
+                ],
+            ],
         ]  # [chapter][map][node(str)] (lst["", enemies])
         self.__dict__.update(
-            yaml_to_dict(os.path.join(MAP_ROOT, "decisive_battle", f"3.yaml"))
+            yaml_to_dict(os.path.join(MAP_ROOT, 'decisive_battle', '3.yaml')),
         )
         self.score = 10
         self.level = 1  # 副官等级
@@ -83,10 +77,10 @@ class DecisiveStats:
     def next(self):
         if self.node == self.map_end[self.chapter][self.map]:
             self.map += 1
-            self.node = "A"
-            return "quit" if (self.map == 4) else "next"
+            self.node = 'A'
+            return 'quit' if (self.map == 4) else 'next'
         self.node = chr(ord(self.node) + 1)
-        return "continue"
+        return 'continue'
 
     @property
     def enemy_now(self):
@@ -98,7 +92,7 @@ class DecisiveStats:
         self.chapter = chapter
 
     def is_begin(self):
-        return self.node == "A" and self.map == 1
+        return self.node == 'A' and self.map == 1
 
 
 class Logic:
@@ -106,21 +100,19 @@ class Logic:
 
     def __init__(
         self,
-        timer,
+        timer: Timer,
         stats: DecisiveStats,
         level1: list,
         level2: list,
         flagship_priority: list,
-    ):
+    ) -> None:
         self.timer = timer
         self.config = timer.config
         self.logger = timer.logger
 
         self.level1 = list(set(level1))
-        self.level2 = list(
-            set(["长跑训练", "肌肉记忆"] + self.level1 + level2 + ["黑科技"])
-        )
-        self.flag_ships = list(set(flagship_priority))
+        self.level2 = list({'长跑训练', '肌肉记忆', *self.level1, *level2, '黑科技'})
+        self.flag_ships = list(flagship_priority)
         self.stats = stats
 
     def _choose_ship(self, must=False):
@@ -132,28 +124,26 @@ class Logic:
             choose = [element for element in self.level2 if is_ship(element)]
         else:
             lim = score
-            choose = self.level1 + [
-                element for element in self.level2 if not is_ship(element)
-            ]
+            choose = self.level1 + [element for element in self.level2 if not is_ship(element)]
         result = []
         for target in choose:
-            if target in self.stats.selections.keys():
+            if target in self.stats.selections:
                 cost = self.stats.selections[target][0]
                 if score >= cost and cost <= lim:
                     score -= cost
                     result.append(target)
         if not result and must:
-            result.append(list(self.stats.selections.keys())[0])
+            result.append(next(iter(self.stats.selections.keys())))
         return result
 
     def _use_skill(self):
-        return 3 if (self.stats.node == "A") else 0
+        return 3 if (self.stats.node == 'A') else 0
 
     def need_repair(self):
-        return bool((1 in self.stats.ship_stats or 2 in self.stats.ship_stats))
+        return bool(1 in self.stats.ship_stats or 2 in self.stats.ship_stats)
 
     def _up_level(self):
-        return bool((self.stats.need - self.stats.exp <= 5 and self.stats.score >= 5))
+        return bool(self.stats.need - self.stats.exp <= 5 and self.stats.score >= 5)
 
     def formation(self):
         pass
@@ -163,9 +153,9 @@ class Logic:
 
     def get_best_fleet(self):
         ships = self.stats.ships
-        self.logger.debug(f"拥有舰船: {ships}")
+        self.logger.debug(f'拥有舰船: {ships}')
         best_ships = [
-            "",
+            '',
         ]
         for ship in self.level1:
             if ship not in ships or len(best_ships) == 7:
@@ -184,12 +174,12 @@ class Logic:
             break
 
         for _ in range(len(best_ships), 7):
-            best_ships.append("")
-        self.logger.debug(f"当前最优：{best_ships}")
+            best_ships.append('')  # noqa: PERF401
+        self.logger.debug(f'当前最优：{best_ships}')
         return best_ships
 
     def _retreat(self):
-        return bool((count_ship(self.get_best_fleet()) < 2))
+        return bool(count_ship(self.get_best_fleet()) < 2)
 
     def _leave(self):
         return False
@@ -214,15 +204,15 @@ class DecisiveBattle:
         self,
         timer: Timer,
         chapter=6,
-        level1=["鲃鱼", "U-1206", "U-47", "射水鱼", "U-96", "U-1405"],
-        level2=["U-81", "大青花鱼"],
-        flagship_priority=["U-1405", "U-47", "U-96", "U-1206"],
+        level1=None,
+        level2=None,
+        flagship_priority=None,
         logic=None,
         repair_level=1,
         full_destroy=False,
         *args,
         **kwargs,
-    ):
+    ) -> None:
         """初始化控制模块
         Important Information:
             : 决战逻辑相当优秀, 不建议自行改写, 满配情况约需要 3~6 快速修复打完一轮 E6, 耗时约为 25mins
@@ -240,6 +230,13 @@ class DecisiveBattle:
             : 若当前决战进度为还没打 6-2-A, 则应填写 chapter=6, map=1, node='A'
             : 可以支持断点开始
         """
+        if level1 is None:
+            level1 = ['鲃鱼', 'U-1206', 'U-47', '射水鱼', 'U-96', 'U-1405']
+        if level2 is None:
+            level2 = ['U-81', '大青花鱼']
+        if flagship_priority is None:
+            flagship_priority = ['U-1405', 'U-47', 'U-96', 'U-1206']
+
         self.timer = timer
         self.config = timer.config
         self.repair_strategy = repair_level
@@ -248,13 +245,17 @@ class DecisiveBattle:
         self.stats = DecisiveStats(timer, chapter)
         if logic is None:
             self.logic = Logic(
-                self.timer, self.stats, level1, level2, flagship_priority
+                self.timer,
+                self.stats,
+                level1,
+                level2,
+                flagship_priority,
             )
         self.__dict__.update(kwargs)
 
-    def buy_ticket(self, use="steel", times=3):
+    def buy_ticket(self, use='steel', times=3):
         self.enter_decisive_battle()
-        position = {"oil": 184, "ammo": 235, "steel": 279, "aluminum": 321}
+        position = {'oil': 184, 'ammo': 235, 'steel': 279, 'aluminum': 321}
         self.timer.click(458 * 0.75, 665 * 0.75, delay=1.5)
         self.timer.click(638, position[use], delay=1, times=times)
         self.timer.click(488, 405)
@@ -268,44 +269,50 @@ class DecisiveBattle:
         check_point = CHECK_POINT[self.stats.chapter]
         for i, point in enumerate(check_point):
             if not self.timer.check_pixel(
-                (int(point[0] * 960), int(point[1] * 540)), (250, 244, 253), 30, True
+                (int(point[0] * 960), int(point[1] * 540)),
+                (250, 244, 253),
+                30,
+                True,
             ):
-                self.timer.logger.info(f"识别决战地图参数, 第 {i} 小节正在进行")
+                self.timer.logger.info(f'识别决战地图参数, 第 {i} 小节正在进行')
                 return i
-        self.timer.logger.info(f"识别决战地图参数, 第 3 小节正在进行")
+        self.timer.logger.info('识别决战地图参数, 第 3 小节正在进行')
         return 3
 
     def recognize_node(self, retry=0):
         position = self.timer.wait_images_position(
-            IMG.fight_image[18:20], confidence=0.7
+            IMG.fight_image[18:20],
+            confidence=0.7,
         )
-        self.timer.logger.debug(f"Ship_ICON position: {position}")
+        self.timer.logger.debug(f'Ship_ICON position: {position}')
         cropped_image = crop_rectangle_relative(
-            self.timer.get_screen(), position[0] / 960 - 0.03, 0, 0.042, 1
+            self.timer.get_screen(),
+            position[0] / 960 - 0.03,
+            0,
+            0.042,
+            1,
         )
-        image_path = os.path.join(TUNNEL_ROOT, "1.PNG")
-        cv2.imencode(".PNG", cropped_image)[1].tofile(image_path)
-        processor = os.path.join(TUNNEL_ROOT, "process_recognize_map.exe")
+        image_path = os.path.join(TUNNEL_ROOT, '1.PNG')
+        cv2.imencode('.PNG', cropped_image)[1].tofile(image_path)
+        processor = os.path.join(TUNNEL_ROOT, 'process_recognize_map.exe')
         subprocess.run([processor, image_path])
         # cropped_image = cv2.imread(image_path)
-        with open(os.path.join(TUNNEL_ROOT, "1.out"), mode="r") as f:
+        with open(os.path.join(TUNNEL_ROOT, '1.out')) as f:
             result = f.read()
-            if result != "":
-                self.timer.logger.info(f"识别决战地图参数, 第 {result[0]} 节点正在进行")
+            if result != '':
+                self.timer.logger.info(f'识别决战地图参数, 第 {result[0]} 节点正在进行')
                 return result[0]
-            else:
-                if retry > 3:
-                    self.timer.logger.warning("识别决战地图参数失败, 退出逻辑")
-                    raise BaseException()
-                else:
-                    self.timer.logger.warning(
-                        f"识别决战地图参数失败, 正在重试第 {retry + 1} 次"
-                    )
-                    return self.recognize_node(retry + 1)
+            if retry > 3:
+                self.timer.logger.warning('识别决战地图参数失败, 退出逻辑')
+                raise BaseException
+            self.timer.logger.warning(
+                f'识别决战地图参数失败, 正在重试第 {retry + 1} 次',
+            )
+            return self.recognize_node(retry + 1)
         # result = recognize(cropped_image, "ABCDEFGHIJK")
         # return result[0][1]
 
-    def detect(self, type="enter_map"):
+    def detect(self, type='enter_map'):
         """检查当前关卡状态
         Args:
             type:
@@ -316,51 +323,52 @@ class DecisiveBattle:
             str: ['challenging', 'refreshed', 'refresh']
             str: ['fight_prepare', 'map']
         """
-        if type == "enter_map":
-            _res = ["cant_fight", "challenging", "refreshed", "refresh"]
+        if type == 'enter_map':
+            _res = ['cant_fight', 'challenging', 'refreshed', 'refresh']
             res = self.timer.wait_images(
-                IMG.decisive_battle_image[3:7], after_get_delay=0.2
+                IMG.decisive_battle_image[3:7],
+                after_get_delay=0.2,
             )
-        if type == "running":
-            _res = ["map", "fight_prepare"]
+        if type == 'running':
+            _res = ['map', 'fight_prepare']
             res = self.timer.wait_images(
-                [IMG.decisive_battle_image[1]]
-                + IMG.identify_images["fight_prepare_page"],
+                [IMG.decisive_battle_image[1]] + IMG.identify_images['fight_prepare_page'],
                 gap=0.03,
                 after_get_delay=0.2,
             )
         return _res[res]
 
     def _go_map_page(self):
-        if self.detect("running") == "fight_prepare":
+        if self.detect('running') == 'fight_prepare':
             self.timer.click(30, 30)
             self.timer.wait_image(IMG.decisive_battle_image[1])
 
     def go_fleet_page(self):
-        if self.detect("running") == "map":
+        if self.detect('running') == 'map':
             self.timer.click(900 * 0.75, 667 * 0.75)
             try:
                 self.timer.wait_images(
-                    IMG.identify_images["fight_prepare_page"],
+                    IMG.identify_images['fight_prepare_page'],
                     timeout=5,
                     after_get_delay=1,
                 )
             except:
-                self.timer.logger.warning("进入出征准备页面失败，正在重试")
+                self.timer.logger.warning('进入出征准备页面失败，正在重试')
                 self.go_fleet_page()
 
     def repair(self):
         self.go_fleet_page()
         quick_repair(
-            self.timer, self.repair_strategy
+            self.timer,
+            self.repair_strategy,
         )  # TODO：我的中破比很高，先改成只修大破控制一下用桶
         # quick_repair(self.timer, 2)
 
     def next(self):
         res = self.stats.next()
-        if res in ["next", "quit"]:
-            self.timer.ConfirmOperation(timeout=5, must_confirm=1)  # 确认通关
-            self.timer.ConfirmOperation(timeout=5, must_confirm=1)  # 确认领取奖励
+        if res in ['next', 'quit']:
+            self.timer.confirm_operation(timeout=5, must_confirm=1)  # 确认通关
+            self.timer.confirm_operation(timeout=5, must_confirm=1)  # 确认领取奖励
             get_ship(self.timer)
         return res
 
@@ -392,12 +400,12 @@ class DecisiveBattle:
             )[1]
         except:
             # TODO: 提高OCR对单个数字的识别率
-            self.timer.logger.warning("读取当前可用费用失败")
+            self.timer.logger.warning('读取当前可用费用失败')
             self.stats.score = 0
-        self.timer.logger.debug(f"当前可用费用为：{self.stats.score}")
+        self.timer.logger.debug(f'当前可用费用为：{self.stats.score}')
         results = self.timer.recognize_number(
             crop_image(screen, *COST_AREA),
-            extra_chars="x",
+            extra_chars='x',
             multiple=True,
         )
         costs = [t[1] for t in results]
@@ -407,48 +415,51 @@ class DecisiveBattle:
                 if cost > self.stats.score:
                     continue
             except Exception as e:
-                self.timer.logger.warning(f"读取购买费用出错，错误如下:\n {e}")
+                self.timer.logger.warning(f'读取购买费用出错，错误如下:\n {e}')
                 continue
             ships.append(
                 self.timer.recognize(
                     crop_image(
-                        screen, (SHIP_X[i][0], SHIP_Y[0]), (SHIP_X[i][1], SHIP_Y[1])
+                        screen,
+                        (SHIP_X[i][0], SHIP_Y[0]),
+                        (SHIP_X[i][1], SHIP_Y[1]),
                     ),
                     candidates=self.timer.ship_names,
-                )[1]
+                )[1],
             )
             _costs.append(cost)
             real_position.append(i)
         # print("Scan result:", costs)
         costs = _costs
         selections = {
-            ships[i]: (costs[i], (CHOOSE_X[real_position[i]], CHOOSE_Y))
-            for i in range(len(costs))
+            ships[i]: (costs[i], (CHOOSE_X[real_position[i]], CHOOSE_Y)) for i in range(len(costs))
         }
         if rec_only:
-            return
+            return None
         # ==================做出决策===================
         self.stats.selections = selections
-        self.timer.logger.debug("可购买舰船：", selections)
+        self.timer.logger.debug('可购买舰船：', selections)
         choose = self.logic._choose_ship(
-            must=(self.stats.map == 1 and self.stats.node == "A" and refreshed == True)
+            must=(self.stats.map == 1 and self.stats.node == 'A' and refreshed),
         )
-        if len(choose) == 0 and refreshed == False:
+        if len(choose) == 0 and not refreshed:
             self.timer.click(380, 500)  # 刷新备选舰船
             self.timer.wait_image(
-                [IMG.decisive_battle_image[2], IMG.decisive_battle_image[8]], timeout=2
+                [IMG.decisive_battle_image[2], IMG.decisive_battle_image[8]],
+                timeout=2,
             )
-            self.timer.logger.info("刷新备选舰船")
+            self.timer.logger.info('刷新备选舰船')
             return self.choose(True)
 
         for target in choose:
             cost, p = selections[target]
             self.stats.score -= cost
-            self.timer.logger.debug(f"选择购买：{target}，花费：{cost}，点击位置：{p}")
+            self.timer.logger.debug(f'选择购买：{target}，花费：{cost}，点击位置：{p}')
             self.timer.relative_click(*p)
             if is_ship(target):
                 self.stats.ships.add(target)
         self.timer.click(580, 500)  # 关闭/确定
+        return None
 
     def up_level_assistant(self):
         self.timer.click(75, 667 * 0.75)
@@ -466,7 +477,7 @@ class DecisiveBattle:
                 multiple=True,
             )
             ships = [ship[1] for ship in ship_results]
-            self.timer.logger.info(f"使用技能获得: {ships}")
+            self.timer.logger.info(f'使用技能获得: {ships}')
             for ship in ships:
                 self.stats.ships.add(ship)
         self.timer.relative_click(*SKILL_POS, times=2, delay=0.3)
@@ -479,7 +490,7 @@ class DecisiveBattle:
         CHAPTER_AREA = ((0.818, 0.867), (0.875, 0.81))
         text = self.timer.recognize(
             crop_image(self.timer.get_screen(), *CHAPTER_AREA),
-            allowlist="Ex-0123456789",
+            allowlist='Ex-0123456789',
             # rgb_select=(247, 221, 82),
             tolerance=50,
         )[1]
@@ -495,31 +506,33 @@ class DecisiveBattle:
         self._move_chapter()
 
     def enter_decisive_battle(self):
-        self.timer.goto_game_page("decisive_battle_entrance")
+        self.timer.goto_game_page('decisive_battle_entrance')
         self.timer.click(115, 113, delay=1.5)
         self.detect()
 
     def enter_map(self, check_map=True):
-        self.stats.node = "A"
+        self.stats.node = 'A'
         if check_map:
             self.enter_decisive_battle()
             self._move_chapter()
             stats = self.detect()
             self.stats.map = self.recognize_map()
-            if stats == "refresh":
+            if stats == 'refresh':
                 entered = self.reset_chapter()
                 if entered:
-                    return
-                stats = "refreshed"
-            if stats == "refreshed":
+                    return None
+                stats = 'refreshed'
+            if stats == 'refreshed':
                 # 选用上一次的舰船并进入
                 if self.check_dock_full():
-                    return "full_destroy_success"
+                    return 'full_destroy_success'
                 self.timer.click(500, 500, delay=0.25)
                 self.stats.map = 1
-                for i in range(5):
+                for i in range(5):  # noqa: B007
                     self.timer.click_image(
-                        IMG.decisive_battle_image[7], timeout=12, must_click=True
+                        IMG.decisive_battle_image[7],
+                        timeout=12,
+                        must_click=True,
                     )
                     self.timer.click(873, 500)
                     if (
@@ -535,7 +548,7 @@ class DecisiveBattle:
                     ):
                         break
                 if i > 3:
-                    raise TimeoutError("选择决战舰船失败")
+                    raise TimeoutError('选择决战舰船失败')
             else:
                 self.timer.click(500, 500, delay=0)
         else:
@@ -544,7 +557,7 @@ class DecisiveBattle:
             self.timer.click(500, 500, delay=0)
 
         if self.check_dock_full():
-            return "full_destroy_success"
+            return 'full_destroy_success'
 
         res = self.timer.wait_images(
             [IMG.decisive_battle_image[1], IMG.decisive_battle_image[3]],
@@ -553,7 +566,7 @@ class DecisiveBattle:
         )
         if res is None:
             raise ImageNotFoundErr("Can't Identify on enter_map")
-        return "other chapter is running" if (res == 1) else "ok"
+        return 'other chapter is running' if (res == 1) else 'ok'
 
     def check_dock_full(self):
         """
@@ -562,12 +575,11 @@ class DecisiveBattle:
         if self.timer.wait_images(IMG.symbol_image[12], timeout=2) is not None:
             if self.full_destroy:
                 self.timer.relative_click(0.38, 0.565)
-                DestroyShip(self.timer)
+                destroy_ship(self.timer)
                 self.enter_map()
                 return True
-            else:
-                self.timer.logger.warning("船舱已满, 但不允许解装")
-                return False
+            self.timer.logger.warning('船舱已满, 但不允许解装')
+            return False
         return False
 
     def retreat(self):
@@ -582,35 +594,36 @@ class DecisiveBattle:
             self.stats.need = 20
             src = self.timer.recognize(
                 crop_image(self.timer.get_screen(), *EXP_AREA),
-                allowlist="Lv.(/)0123456789",
+                allowlist='Lv.(/)0123456789',
             )[1]
             try:
-                index1 = src.index("(")
+                index1 = src.index('(')
             except ValueError:
-                index1 = float("inf")  # 如果没有找到，设置为无穷大
+                index1 = float('inf')  # 如果没有找到，设置为无穷大
             try:
-                index2 = src.index("（")
+                index2 = src.index('（')
             except ValueError:
-                index2 = float("inf")  # 如果没有找到，设置为无穷大
+                index2 = float('inf')  # 如果没有找到，设置为无穷大
             i1 = min(index1, index2)
-            if i1 == float("inf"):
+            if i1 == float('inf'):
                 raise ValueError("未找到 '(' 或 '（'")
-            i2 = src.index("/")
-            src = src.rstrip(")）")
+            i2 = src.index('/')
+            src = src.rstrip(')）')
             self.stats.exp = int(src[i1 + 1 : i2])
             self.stats.need = int(src[i2 + 1 :])
             self.timer.logger.debug(
-                f"当前经验：{self.stats.exp}，升级需要经验：{self.stats.need}"
+                f'当前经验：{self.stats.exp}，升级需要经验：{self.stats.need}',
             )
         except:
-            self.timer.logger.warning("识别副官升级经验数值失败")
+            self.timer.logger.warning('识别副官升级经验数值失败')
 
     def _before_fight(self):
-        if self.timer.wait_image(IMG.confirm_image[1:], timeout=1) != False:
+        if self.timer.wait_image(IMG.confirm_image[1:], timeout=1):
             self.timer.click(300, 225)  # 选上中下路
-            self.timer.ConfirmOperation(must_confirm=1)
+            self.timer.confirm_operation(must_confirm=1)
         if self.timer.wait_image(
-            [IMG.decisive_battle_image[2], IMG.decisive_battle_image[8]], timeout=2
+            [IMG.decisive_battle_image[2], IMG.decisive_battle_image[8]],
+            timeout=2,
         ):
             self.choose()  # 获取战备舰队
         self._get_exp()
@@ -628,15 +641,16 @@ class DecisiveBattle:
         _fleet = self.logic.get_best_fleet()
         if self.logic._retreat():
             self.retreat()
-            return "retreat"
+            return 'retreat'
         if self.logic._leave():
             self.leave()
-            return "leave"
+            return 'leave'
         if self.stats.fleet != _fleet:
             self._change_fleet(_fleet)
             self.stats.ship_stats = detect_ship_stats(self.timer)
         if self.logic.need_repair():
             self.repair()
+        return None
 
     def _after_fight(self):
         self.timer.logger.info(self.stats.ship_stats)
@@ -655,26 +669,25 @@ class DecisiveBattle:
             gap=0,
             timeout=16,
         )
-        if res == None:
-            raise TimeoutError("选择舰船时点击超时")
+        if res is None:
+            raise TimeoutError('选择舰船时点击超时')
         ships = self.timer.recognize_ship(
-            self.timer.get_screen()[:, :1048], self.timer.ship_names
+            self.timer.get_screen()[:, :1048],
+            self.timer.ship_names,
         )
-        self.timer.logger.info(f"其它可用舰船：{[item[1] for item in ships]}")
+        self.timer.logger.info(f'其它可用舰船：{[item[1] for item in ships]}')
         for ship in ships:
             self.stats.ships.add(ship[1])
         self.timer.relative_click(0.05, 0.05)
         res = self.timer.wait_images(
-            IMG.identify_images["fight_prepare_page"],
+            IMG.identify_images['fight_prepare_page'],
             gap=0.03,
             after_get_delay=0.2,
         )
 
     def _during_fight(self):
         formation = get_formation(self.stats.fleet, self.stats.enemy_now)
-        night = (
-            self.stats.node in self.stats.key_points[self.stats.chapter][self.stats.map]
-        )
+        night = self.stats.node in self.stats.key_points[self.stats.chapter][self.stats.map]
         plan = DecisiveBattlePlan(
             self.timer,
             formation=formation,
@@ -682,9 +695,7 @@ class DecisiveBattle:
             ship_stats=self.stats.ship_stats,
         )
         plan.run()
-        self.stats.ship_stats = plan.Info.fight_history.get_fight_results()[
-            -1
-        ].ship_stats
+        self.stats.ship_stats = plan.info.fight_history.get_fight_results()[-1].ship_stats
 
     def _change_fleet(self, fleet):
         self.go_fleet_page()
@@ -703,7 +714,7 @@ class DecisiveBattle:
         #         self._reset()
         #         return 'continue'
 
-        if res == "retreat":
+        if res == 'retreat':
             self.enter_map(check_map=False)
             self._reset()
             return self.fight()
@@ -715,12 +726,12 @@ class DecisiveBattle:
         self.enter_map()
         while True:
             res = self.fight()
-            if res == "quit":
+            if res == 'quit':
                 return
-            elif res == "next":
+            if res == 'next':
                 self.enter_map(False)
 
-    def reset_chapter(self):
+    def reset_chapter(self) -> None | Literal[True]:
         """使用磁盘重置关卡, 并重置状态"""
         # Todo: 缺少磁盘报错
         self._reset()
@@ -728,18 +739,18 @@ class DecisiveBattle:
         self.timer.relative_click(0.5, 0.925)
         if self.check_dock_full():
             return True
-        else:
-            self.timer.ConfirmOperation(must_confirm=True)
+        self.timer.confirm_operation(must_confirm=True)
+        return None
 
     def _reset(self):
         self.stats.reset()
 
 
 class DecisiveBattlePlan(BattlePlan):
-    def __init__(self, timer: Timer, formation, night, ship_stats):
+    def __init__(self, timer: Timer, formation, night, ship_stats) -> None:
         super().__init__(timer, None)
-        self.Info = DecisiveBattleInfo(timer)
-        self.Info.ship_stats = ship_stats
+        self.info = DecisiveBattleInfo(timer)
+        self.info.ship_stats = ship_stats
         self.node.formation = formation
         self.node.night = night
 
@@ -748,8 +759,8 @@ class DecisiveBattlePlan(BattlePlan):
 
 
 class DecisiveBattleInfo(BattleInfo):
-    def __init__(self, timer: Timer):
+    def __init__(self, timer: Timer) -> None:
         super().__init__(timer)
-        self.end_page = "unknown_page"
-        self.state2image["battle_page"] = [IMG.decisive_battle_image[1], 5]
-        self.state2image["result"] = [IMG.fight_image[3], 60, 0.7]
+        self.end_page = 'unknown_page'
+        self.state2image['battle_page'] = [IMG.decisive_battle_image[1], 5]
+        self.state2image['result'] = [IMG.fight_image[3], 60, 0.7]
