@@ -1,23 +1,22 @@
-import os
 import time
+from typing import Any
 
 from autowsgr.constants.custom_exceptions import ImageNotFoundErr
-from autowsgr.constants.data_roots import IMG_ROOT
 from autowsgr.constants.image_templates import IMG
 from autowsgr.timer import Timer
-from autowsgr.utils.math_functions import CalcDis
+from autowsgr.utils.math_functions import cal_dis
 
 
 class Event:
-    def __init__(self, timer: Timer, event_name):
+    def __init__(self, timer: Timer, event_name: str) -> None:
         self.timer = timer
         self.logger = timer.logger
 
         self.event_image = IMG.event[event_name]
-        self.common_image = IMG.event["common"]
-        self.enemy_image = IMG.event["enemy"]
+        self.common_image = IMG.event['common']
+        self.enemy_image = IMG.event['enemy']
 
-        self.common_image["monster"] = [
+        self.common_image['monster'] = [
             self.common_image.little_monster,
             self.common_image.big_monster,
         ]
@@ -35,30 +34,29 @@ class Event:
         """
         res = self.timer.wait_images(self.common_image.hard + self.common_image.easy)
         if res is None:
-            self.logger.error("ImageNotFoundErr: difficulty image not found")
+            self.logger.error('ImageNotFoundErr: difficulty image not found')
             if self.timer.wait_image(self.event_image[2]):
                 self.logger.info(
-                    "成功进入活动页面，未检测到切换难度图标，请检查是否通关简单难度"
+                    '成功进入活动页面，未检测到切换难度图标，请检查是否通关简单难度',
                 )
                 return 0
             self.timer.log_screen()
-            raise ImageNotFoundErr()
+            raise ImageNotFoundErr
 
         if self.timer.image_exist(self.common_image.hard, need_screen_shot=False):
             return 0
-        else:
-            return 1
+        return 1
 
-    def change_difficulty(self, chapter, retry=True):
-        r_difficulty = int(chapter in "Hh")
+    def change_difficulty(self, chapter, retry=True) -> Any | None:
+        r_difficulty = int(chapter in 'Hh')
         difficulty = self.get_difficulty()
 
         if r_difficulty != difficulty:
             time.sleep(0.2)
-            if int(chapter in "Hh"):
+            if int(chapter in 'Hh'):
                 if not self.timer.click_image(self.common_image.hard):
-                    self.logger.error("请检查是否通关简单难度")
-                    raise ImageNotFoundErr("Can't change difficulty")
+                    self.logger.error('请检查是否通关简单难度')
+                    raise ImageNotFoundErr
             else:
                 self.timer.click_image(self.common_image.easy)
 
@@ -66,13 +64,14 @@ class Event:
                 if retry:
                     return self.change_difficulty(chapter, False)
                 self.timer.log_screen()
-                raise ImageNotFoundErr("Can't change difficulty")
+                raise ImageNotFoundErr
+        return None
 
 
 class PatrollingEvent(Event):
     """巡戈作战活动 类"""
 
-    def __init__(self, timer: Timer, event_name, map_positions):
+    def __init__(self, timer: Timer, event_name, map_positions) -> None:
         """
         Args:
             map_positions : 从主页面点进活动后, 去到对应地图需要点的位置
@@ -87,7 +86,7 @@ class PatrollingEvent(Event):
 
     def enter_map(self, chapter, map):
         """从活动地图选择界面进入到巡游地图"""
-        assert chapter in "HEhe"
+        assert chapter in 'HEhe'
         assert map in range(1, 7)
         self.change_difficulty(chapter)
         if map <= 3:
@@ -97,19 +96,14 @@ class PatrollingEvent(Event):
             self.timer.swipe(600, 300, 100, 300, duration=0.4, delay=0.15)
             self.timer.swipe(600, 300, 100, 300, duration=0.4, delay=0.15)
         self.timer.click(*self.MAP_POSITIONS[map], delay=0.25)
-        assert (
-            self.timer.wait_image(self.event_image[2]) is not False
-        )  # 是否成功进入地图
+        assert self.timer.wait_image(self.event_image[2]) is not False  # 是否成功进入地图
 
     def go_fight_prepare_page(self):
         self.timer.click(789, 455)
-        assert (
-            self.timer.wait_image(IMG.identify_images["fight_prepare_page"])
-            is not False
-        )
+        assert self.timer.wait_image(IMG.identify_images['fight_prepare_page']) is not False
 
     def random_walk(self):
-        "随机游走,寻找敌人"
+        """随机游走,寻找敌人"""
         ways = ((0, 1), (0, -1), (1, 0), (-1, 0), (1, 1), (1, -1), (-1, 1), (-1, -1))
         import random
 
@@ -126,26 +120,25 @@ class PatrollingEvent(Event):
     def get_close(self, images):
         while True:
             ret = self.timer.wait_images_position(
-                images, confidence=0.8, gap=0.03, timeout=1
+                images,
+                confidence=0.8,
+                gap=0.03,
+                timeout=1,
             )
-            if (
-                CalcDis([ret[0]], [480]) ** 0.5 < 320
-                and CalcDis([ret[1]], [270]) ** 0.5 < 180
-            ):
+            if cal_dis([ret[0]], [480]) ** 0.5 < 320 and cal_dis([ret[1]], [270]) ** 0.5 < 180:
                 return ret
-            if ret[0] > 480:
-                ret = (ret[0] - 130, ret[1])
-            else:
-                ret = (ret[0] + 130, ret[1])
+            ret = (ret[0] - 130, ret[1]) if ret[0] > 480 else (ret[0] + 130, ret[1])
             self.timer.click(*ret)
 
     def find(self, images, max_times=20):
         for _ in range(max_times):
             ret = self.timer.wait_images_position(
-                images, confidence=0.75, gap=0.03, timeout=1
+                images,
+                confidence=0.75,
+                gap=0.03,
+                timeout=1,
             )
             if ret is not None:
                 return ret
-            else:
-                self.random_walk()
+            self.random_walk()
         return None
