@@ -1,1 +1,1 @@
-from .ocr_backend import EasyocrBackend, OCRBackend, PaddleOCRBackend
+from autowsgr.timer.backends.ocr_backend import EasyocrBackend, OCRBackend, PaddleOCRBackend
