@@ -1,2 +1,2 @@
-from .android_controller import AndroidController
-from .windows_controller import WindowsController
+from autowsgr.timer.controllers.android_controller import AndroidController
+from autowsgr.timer.controllers.windows_controller import WindowsController
