@@ -4,9 +4,10 @@ import time
 from autowsgr.game.build import BuildManager
 from autowsgr.scripts.main import start_script
 
+
 resources = [90, 30, 90, 30]
 
-timer = start_script(f"{os.path.dirname(os.path.abspath(__file__))}/user_settings.yaml")
+timer = start_script(f'{os.path.dirname(os.path.abspath(__file__))}/user_settings.yaml')
 build_manager = BuildManager(timer)
 
 while build_manager.has_empty_slot():
