@@ -1,10 +1,12 @@
 import os
 import sys
 
+
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from autowsgr.constants.other_constants import ALL_SHIP_TYPES_CN, CN_TYPE_TO_EN_TYPE
 from autowsgr.ocr.ship_name import get_allow, recognize
+
 
 DEBUG = False
 
@@ -25,8 +27,8 @@ def split_str(str, keywords):
 
 def recognize_decisive_enemy(image):
     allow = get_allow(ALL_SHIP_TYPES_CN)
-    char = "A"
-    for box in recognize(image, char_list=allow + "/"):
+    char = 'A'
+    for box in recognize(image, char_list=allow + '/'):
         try:
             text = box[1]
             if DEBUG:
@@ -35,14 +37,14 @@ def recognize_decisive_enemy(image):
             enemy = split_str(text, keywords=ALL_SHIP_TYPES_CN)
             for i in range(len(enemy)):
                 enemy[i] = CN_TYPE_TO_EN_TYPE[enemy[i]]
-            print(char, ":", [""] + enemy)
+            print(f'{char} : {enemy}')
         except Exception as e:
             print(e)
             print(char)
         char = chr(ord(char) + 1)
 
 
-if __name__ == "__main__":
-    print("Please Input The Image File:")
+if __name__ == '__main__':
+    print('Please Input The Image File:')
     path = input()
     recognize_decisive_enemy(path)
